Arquivos Iniciais

+++ resources/views/admin/informativo/index.blade.php
@@ -14,9 +14,11 @@
                 <div class="card">
                     <div class="card-header">
                         <i class="fa fa-align-justify"></i> {{ trans('admin.informativo.actions.index') }}
-                        <a class="btn btn-primary btn-spinner btn-sm pull-right m-b-0"
-                           href="{{ url('admin/informativos/create') }}" role="button"><i
-                                class="fa fa-plus"></i>&nbsp; {{ trans('admin.informativo.actions.create') }}</a>
+                        @if(auth()->user()->is_admin === 1)
+                            <a class="btn btn-primary btn-spinner btn-sm pull-right m-b-0"
+                               href="{{ url('admin/informativos/create') }}" role="button"><i
+                                    class="fa fa-plus"></i>&nbsp; {{ trans('admin.informativo.actions.create') }}</a>
+                        @endif
                     </div>
                     <div class="card-body" v-cloak>
                         <div class="card-block">
@@ -128,16 +130,25 @@
                                     <td>
                                         <div class="row no-gutters">
                                             <div class="col-auto">
-                                                <a class="btn btn-sm btn-spinner btn-info"
-                                                   :href="item.resource_url + '/edit'"
+                                                <a class="btn btn-sm btn-spinner btn-success"
+                                                   :href="item.resource_url"
                                                    title="{{ trans('brackets/admin-ui::admin.btn.edit') }}"
-                                                   role="button"><i class="fa fa-edit"></i></a>
+                                                   role="button"><i class="fa fa-eye"></i></a>
                                             </div>
-                                            <form class="col" @submit.prevent="deleteItem(item.resource_url)">
-                                                <button type="submit" class="btn btn-sm btn-danger"
-                                                        title="{{ trans('brackets/admin-ui::admin.btn.delete') }}"><i
-                                                        class="fa fa-trash-o"></i></button>
-                                            </form>
+                                            @if(auth()->user()->is_admin === 1)
+                                                <div class="col-auto">
+                                                    <a class="btn btn-sm btn-spinner btn-info"
+                                                       :href="item.resource_url + '/edit'"
+                                                       title="{{ trans('brackets/admin-ui::admin.btn.edit') }}"
+                                                       role="button"><i class="fa fa-edit"></i></a>
+                                                </div>
+                                                <form class="col" @submit.prevent="deleteItem(item.resource_url)">
+                                                    <button type="submit" class="btn btn-sm btn-danger"
+                                                            title="{{ trans('brackets/admin-ui::admin.btn.delete') }}">
+                                                        <i
+                                                            class="fa fa-trash-o"></i></button>
+                                                </form>
+                                            @endif
                                         </div>
                                     </td>
                                 </tr>
